+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 **ATTN**: This project uses [semantic versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [v0.8.0] - 2022-06-09
 ### Added
 - Added drink Greedy Salvation that cures Zombie Virus.
 
@@ -61,7 +63,8 @@ All notable changes to this project will be documented in this file.
 Slender Doe. 
 - Added EN and RU translations.
 
-[Unreleased]: https://github.com/gorcon/rcon-cli/compare/v0.7.1...HEAD
+[Unreleased]: https://github.com/gorcon/rcon-cli/compare/v0.8.0...HEAD
+[v0.7.1]: https://github.com/gorcon/rcon-cli/compare/v0.7.1...v0.8.0
 [v0.7.1]: https://github.com/gorcon/rcon-cli/compare/v0.7.0...v0.7.1
 [v0.7.0]: https://github.com/gorcon/rcon-cli/compare/v0.6.0...v0.7.0
 [v0.6.0]: https://github.com/gorcon/rcon-cli/compare/v0.5.0...v0.6.0